refactor: Simplify AuthModal forms by extracting sign-in and sign-up logic into custom hooks
